+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.0.0 - Framework Rewrite [2024-11-xx]
+## 3.0.0 - Framework Rewrite [2024-12-xx]
 
 To ease of development for future maintainer or anyone who is much more comfortable with framework rather than raw dog HTML. As much as I prefer working on vanilla side for performance reason, I do want people to contribute without hassle and I find Astro+SolidJS to be solid (ha) choice for performance and ease of development balance.
 
@@ -14,6 +14,9 @@ To ease of development for future maintainer or anyone who is much more comforta
    -  Hip Hop
    -  Collider
    -  Britney Spears
+   -  Monstercat 2.0
+   -  Metallica
+-  Variant color scheme from Metallica
 -  Info color scheme for color picker
 
 ### Changed